requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pandas>=2.2.2
datasets>=2.20.0
transformers>=4.44.0
peft>=0.12.0
accelerate>=0.33.0
bitsandbytes>=0.43.3
